requirements/base.txt
# Base application requirements

django>=4.0.0,<5.0.0
django-ninja
django-cors-headers
pyjwt
pytz